requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorboard>=2.12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.7.0
pillow>=9.5.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0